Replace goog.object.clear() with ol.object.clear()
